+++ src/features/listing/ui/detail/ListingDetail.styles.ts
