+++ src/features/listing/ui/index.ts
@@ -1,8 +1,8 @@
-import FilterSidebar from './FilterSidebar';
-import ListingList from './ListingList';
-import ControlsDesktop from './ControlsDesktop';
-import ControlsMobile from './ControlsMobile';
-import ListingDetail from './ListingDetail';
+import FilterSidebar from './main/filter/FilterSidebar';
+import ListingList from './main/list/ListingList';
+import ControlsDesktop from './main/ControlsDesktop';
+import ControlsMobile from './main/ControlsMobile';
+import ListingDetail from './detail/ListingDetail';
 export {
   FilterSidebar,
   ListingList,
