+++ src/features/listing/ui/main/ControlsDesktop.tsx
@@ -1,9 +1,9 @@
 'use client';
 import styled from 'styled-components';
-import ListingSortDropdown from './ListingSortDropdown';
-import ListingAreaButton from './ListingAreaButton';
-import ListingSearchBar from './ListingSearchBar';
-import ActiveFilterTag from './ActiveFilterTag';
+import ListingSortDropdown from './filter/ListingSortDropdown';
+import ListingAreaButton from './filter/ListingAreaButton';
+import ListingSearchBar from './filter/ListingSearchBar';
+import ActiveFilterTag from './filter/ActiveFilterTag';
 
 const ControlsDesktop = () => {
   return (
